requirements.txt
dj-database-url==0.5.0
Django==3.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
whitenoise==5.2.0
spotipy==2.17.1
pymongo==3.11.3
dnspython==2.1.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
whitenoise==5.2.0
geopy==2.1.0
geojson==2.5.0
beautifulsoup4==4.9.3 
